web/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
tensorflow>=2.15.0
pillow>=10.1.0
numpy>=1.24.3

# NOTA: TensorFlow requiere Python 3.9-3.12
# Si usas Python 3.14, necesitas crear el venv con Python 3.11 o 3.12:
# python3.11 -m venv venv
# o
# python3.12 -m venv venv

